Add InputStep step type and more documentation.
The InputStep will be used to track readiness of inputs to the protocol.
This allows uniform handling of all inputs to all steps, rather than
special-casing inputs not derived as the output of an earlier step.

In a nutshell, when an InputStep needs to be evaluated, the relevant
party will use this as a signal to find the relevant input, do any
necessary data transformations, etc.

This commit also adds an example ExchangeWorkflow textproto to illustrate
a potential protocol.