Update exchange shared input/output to support urls
The size of the blobs exchanged in the api steps is likely to be too
large to reasonably support directly passing via the API. On top of
that, we want to avoid having the Kingdom own the storage for some of
these steps. To avoid issues, this swaps the 'shared' values to string
from blob.

* Only updates the shared input/output map values from blob to string
* Implementation is done in another step.