Use ephemeral sender keys for hybrid encryption.
This means that the sender generates a new asymmetric key pair for each message, and throws away the private key after encrypting. The encryption output contains the public key from this pair along with the ciphertext. See https://en.wikipedia.org/wiki/Integrated_Encryption_Scheme.

For Measurements, this is comprised of the following changes:
* The per-Measurement public key that the MC generates is only used as a recipient key, e.g. for encrypting the result.
  * The MC generates a separate ephemeral key pair for each `RequisitionSpec`.
* The `result_public_key field` is no longer needed, since that key is embedded in the encrypted value.