Add identification to the exchange_workflows
This allows a party to identify which trusted cert should be used to
validate inputs as well as where shared storage is located (while
keeping that location hidden from the Kingdom).

These ids are:

1. A storage salt. To keep the location of shared storage secret from the kingdom (in case bucket security was misconfigured) this provides a salt value unknown to the kingdom (being encrypted) that, when concatenated with the exchange date and the other two values below will make a bucket name that is highly unique and unlikely to have a collision.

Note that the hash output will have to be limited to legal names for S3 and GCS buckets, IE: <64 characters, lowercase letters, numbers, and dashes (but not the first or last character).

2 & 3. the names of the EDP and MP certs in the clientCerts bundle. This follows the pattern of how requisiton.proto does things.

The cert names will allow us grab the right certs for each exchange step attempt from clientCerts to validate the exchange_workflow itself and shared inputs.