Add identification to the exchange_workflows
This allows a party to identify which trusted cert should be used to
validate inputs as well as which key to use when building the shared storage client

These ids are the names of the EDP and MP certs in the clientCerts bundle. This follows the pattern of how requisiton.proto does things.

The cert names will allow us grab the right certs for each exchange step attempt from clientCerts to validate the exchange_workflow itself and shared inputs.