Refactor custom direct methodology (#178)
## TL;DR
This pull request introduces changes to the `direct_computation.proto`
and `protocol_config.proto` files. The changes include the addition of a
new message `FrequencyVariances` and a `oneof` field `variance` in the
`CustomDirectMethodology` message. It also modifies the comment for
`custom_direct_methodology` in `ProtocolConfig`.
> 
## What changed
In `direct_computation.proto`, the `variance` field in
`CustomDirectMethodology` has been replaced with a `oneof` field that
can either be a `scalar` or `FrequencyVariances`. The
`FrequencyVariances` message includes two maps for `variances` and
`k_plus_variances`, which represent the variances of different types of
frequency distributions from frequency 1 to the maximum frequency
specified in the measurement spec.

> 
In `protocol_config.proto`, the comment for `custom_direct_methodology`
has been updated to include a requirement for data providers to
guarantee the independence of the result by salting their hash
functions.
> 
## Why make this change
This change allows for more flexibility in specifying the variance in
the `CustomDirectMethodology` message. It also clarifies the
requirements for data providers who choose to use a custom direct
methodology.