Update FulfillRequisitionRequest and ProtocolConfig for ShuffleBasedSecretSharing protocol (#191)
For shuffle based secret sharing protocol, requisitions are fulfilled by
EDPs with either chunks of data (a blob) or a seed. The seed can be
expanded into a deterministic blob with the same PRNG by workers.
The change is back-compatible.

Also added ProtocolConfig in FulfillRequisitionRequest for this
issue(https://github.com/world-federation-of-advertisers/cross-media-measurement/issues/1329).