docs: Clarify which intervals EnumerateValidModelLines considers
The primary use case of the method is for a MeasurementConsumer caller to determine which ModelLine to specify for a Measurement. For this purpose, only the DataProvider data availability for each ModelLine is relevant. The active range of the ModelLine is only transitively considered by the fact that DataProvider data availability is inherently constrained by the ModelLine active range.